added dog route errors (400) for edge cases

+++ sniffr/dog_routes/dog_routes.py
@@ -24,7 +24,7 @@ def get_dog(dog_id):
 
     else:
         response = {}
-        return jsonify(response)
+        return make_response("Dog not found", 400)
 
 
 # Get All Dogs
@@ -89,20 +89,18 @@ def post_dog(current_user):
 
     # Kick back if actvities list is > 3
     if len(content['activities']) > 3:
-        return jsonify({'message': "Dogs are limited to 3 activities"}, 400)
+        return make_response("Dogs are limited to 3 activities", 400)
     
     # If dog_id not in body then they are trying to create
     # If dog_id in body then updating content
     if "dog_id" in content.keys():
         queried_dog = (
             db.session.query(Dog)
-            .join(DogActivity)
-            .join(Activity)
             .filter(Dog.dog_id == int(content["dog_id"]))
             .filter(Dog.owner_id == user_id)
             .first()
         )
-
+        
         if queried_dog:
             # Update properties
             queried_dog.dog_name = content["dog_name"]
@@ -119,15 +117,16 @@ def post_dog(current_user):
 
             db.session.commit()
 
-            # TODO: Delete dog's old activities
+            # Delete dog's old activities
             old_activities = (
                 db.session.query(DogActivity)
                     .filter(DogActivity.dog_id == int(content["dog_id"]))
                     .all()
                 )
-            for each in old_activities:
-                db.session.delete(each)
-                db.session.commit()
+            if old_activities:
+                for each in old_activities:
+                    db.session.delete(each)
+                    db.session.commit()
 
             # Add dog's activities
             for activity_id in content['activities']:
@@ -136,13 +135,12 @@ def post_dog(current_user):
                 db.session.add(dogs_activity)
                 db.session.commit()
 
-
             response = process_dog(queried_dog)
 
             return jsonify(response)
 
         else:
-            return jsonify({}), 200
+            return make_response("Dog not found", 400)
 
     else:
         # create dog
@@ -164,24 +162,15 @@ def post_dog(current_user):
         db.session.add(new_dog)
         db.session.commit()
 
-        for activity_id in content['activities']:
-            dogs_activity = DogActivity(dog_id=new_dog.dog_id, activity_id=activity_id)
-            db.session.add(dogs_activity)
-            db.session.commit()
+        if content['activities']:
+            for activity_id in content['activities']:
+                dogs_activity = DogActivity(dog_id=new_dog.dog_id, activity_id=activity_id)
+                db.session.add(dogs_activity)
+                db.session.commit()
 
-        queried_dog = (
-            db.session.query(Dog)
-            .join(Breed)
-            .join(User)
-            .join(DogActivity)
-            .join(Activity)
-            .filter(Dog.dog_id == new_dog.dog_id)
-            .first()
-        )
-        
-        response = process_dog(queried_dog)
+        response = process_dog(new_dog)
 
-        return jsonify(response), 201
+        return make_response(jsonify(response), 200)
 
 
 # Delete Dog
@@ -201,7 +190,7 @@ def delete_dog(current_user, dog_id):
             db.session.delete(queried_dog)
             db.session.commit()
 
-            return jsonify({}), 200
+            return make_response(jsonify({}), 200)
 
     else:
-        return jsonify({}), 204
+        return make_response("Dog not found", 400)
